Added AGG to the main AeonGUI project and did some cleanup.

+++ agg/src/agg_gsv_text.cpp
@@ -571,7 +571,8 @@ namespace agg
             m_text = m_chr;
             return;
         }
-        unsigned new_size = strlen(text) + 1;
+		///@todo Consider changing agg::pod_array to use size_t rather than unsigned
+        unsigned new_size = static_cast<unsigned>(strlen(text) + 1);
         if(new_size > m_text_buf.size())
         {
             m_text_buf.resize(new_size);
